requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
streamlit==1.38.0
google-generativeai>=0.8.0
python-multipart==0.0.9
librosa>=0.10.0
soundfile>=0.12.0
scikit-learn>=1.3.0
requests>=2.31.0
